初始化 oxo

+++ plm-bom-configuration-sdk/src/main/java/com/nio/ngfs/plm/bom/configuration/sdk/dto/common/PageData.java
@@ -22,7 +22,7 @@ public class PageData<T> {
     private int pageSize;
 
     /**一共多少条*/
-    private long totalItems;
+    private int totalItems;
     
     /**查询集合*/
     private List<T> dataList;
